+++ src/lingua/extract.py
@@ -17,6 +17,7 @@ from lingua.extractors import register_extractors
 from lingua.extractors.babel import register_babel_plugins
 from lingua.extractors import EXTRACTORS
 from lingua.extractors import EXTENSIONS
+from lingua import lingua_version
 
 
 def po_timestamp():
@@ -134,6 +135,7 @@ def create_catalog(options):
     catalog.metadata['MIME-Version'] = '1.0'
     catalog.metadata['Content-Type'] = 'text/plain; charset=UTF-8'
     catalog.metadata['Content-Transfer-Encoding'] = '8bit'
+    catalog.metadata['Generated-By'] = 'Lingua %s' % lingua_version()
     return catalog